+++ examples/rust/test_calls.rs
@@ -1,0 +1,7 @@
+pub fn init_config_file() {
+    crate::init::init_global_dirs();
+}
+
+pub fn init_global_dirs() {
+    println!("Initializing");
+}